requirements.txt
# Django>=2.2.6,<2.3.0
# graphene-django>=2.0,<2.8.4
# graphene==2.1.8
# django-graphql-jwt>=0.3.0,<0.4
# django-cors-headers>=3.2.1,<3.3
# psycopg2>=2.7.5,<2.8.0

# flake8>=3.6.0,<3.7.0

# sendgrid==6.2.2

aniso8601==7.0.0
autopep8==1.4.4
Django==2.2.10
django-graphql-jwt==0.2.2
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
django-cors-headers>=3.2.1,<3.3
psycopg2>=2.7.5,<2.8.0
flake8>=3.6.0,<3.7.0
graphql-relay==2.0.0
pep8==1.7.1
promise==2.2.1
pycodestyle<2.5.0,>=2.4.0
PyJWT==1.7.1
python-http-client==3.2.1
pytz==2019.3
Rx==1.6.1
sendgrid==6.1.0
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
gunicorn==19.9.0